Separate main stuff from IHM
